added graph document structure
